[Game Events] Using hpp for headers.
