+++ tools/game-events/game-events.vcxproj
@@ -159,7 +159,7 @@
     <ClCompile Include="src\Main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\GameEvents.h" />
+    <ClInclude Include="src\GameEvents.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\libraries\core\core.vcxproj">
