+++ tools/game-events/game-events.vcxproj.filters
@@ -23,7 +23,7 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\GameEvents.h">
+    <ClInclude Include="src\GameEvents.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
