+++ tools/game-events/src/GameEvents.cpp
@@ -1,4 +1,4 @@
-#include "GameEvents.h"
+#include "GameEvents.hpp"
 
 #include <cstdio>
