+++ tools/game-events/src/GameEvents.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+
+#include <cstdint>
+
+#include "core/DetourHook.hpp"
+
+
+class GameEvents
+{
+public:
+    GameEvents();
+    ~GameEvents();
+
+private:
+    void OnGameEvent(const uint8_t* eventData, int32_t eventID, uint32_t eventSize);
+
+private:
+    static void DetourProcessGameEvents();
+
+private:
+    DetourHook m_DetourProcessGameEvents;
+};
